Mejora 3: soporte para decimales

+++ 04_escenario_ramas/EcuacionSegundoGradoRamas/src/Main.java
@@ -2,10 +2,10 @@ import java.util.Scanner;
 public class Main {
  public static void main(String[] args) {
  Scanner sc = new Scanner(System.in);
- int a, b, c;
+double a, b, c;
  double discriminante, x1, x2;
  System.out.print("Ingrese a: ");
- a = sc.nextInt();
+ a = sc.nextDouble();
  System.out.print("Ingrese b: ");
  b = sc.nextInt();
  System.out.print("Ingrese c: ");
